# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FincaGest360 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lotes:
    get:
      operationId: listLotes
      tags: [lotes]
      summary: List all lots
      parameters:
        - name: fincaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of lots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lote"
    post:
      operationId: createLote
      tags: [lotes]
      summary: Create a lot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lote"
  /lotes/{id}:
    put:
      operationId: updateLote
      tags: [lotes]
      summary: Update a lot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoteBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lote"
    delete:
      operationId: deleteLote
      tags: [lotes]
      summary: Delete a lot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /produccion:
    get:
      operationId: listProduccion
      tags: [produccion]
      summary: List production records
      parameters:
        - name: fincaId
          in: query
          schema:
            type: integer
        - name: loteId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produccion"
    post:
      operationId: createProduccion
      tags: [produccion]
      summary: Create production record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduccionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produccion"
  /produccion/{id}:
    delete:
      operationId: deleteProduccion
      tags: [produccion]
      summary: Delete production record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /nomina:
    get:
      operationId: listNomina
      tags: [nomina]
      summary: List payroll records
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NominaRecord"
    post:
      operationId: createNomina
      tags: [nomina]
      summary: Create payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNominaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NominaRecord"
  /nomina/{id}:
    delete:
      operationId: deleteNomina
      tags: [nomina]
      summary: Delete payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /insumos:
    get:
      operationId: listInsumos
      tags: [insumos]
      summary: List inputs/supplies
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insumo"
    post:
      operationId: createInsumo
      tags: [insumos]
      summary: Create input/supply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInsumoBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insumo"
  /insumos/{id}:
    delete:
      operationId: deleteInsumo
      tags: [insumos]
      summary: Delete input/supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jornadas:
    get:
      operationId: listJornadas
      tags: [jornadas]
      summary: List work days
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Jornada"
    post:
      operationId: createJornada
      tags: [jornadas]
      summary: Create work day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJornadaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jornada"
  /jornadas/{id}:
    delete:
      operationId: deleteJornada
      tags: [jornadas]
      summary: Delete work day
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Lote:
      type: object
      properties:
        id:
          type: integer
        fincaId:
          type: integer
        nombre:
          type: string
        hectareas:
          type: number
        cultivo:
          type: string
        estado:
          type: string
      required:
        - id
        - fincaId
        - nombre
        - hectareas
        - cultivo
        - estado
    CreateLoteBody:
      type: object
      properties:
        fincaId:
          type: integer
        nombre:
          type: string
        hectareas:
          type: number
        cultivo:
          type: string
        estado:
          type: string
      required:
        - fincaId
        - nombre
        - hectareas
        - cultivo
        - estado
...
    Produccion:
      type: object
      properties:
        id:
          type: integer
        fincaId:
          type: integer
        loteId:
          type: integer
        fecha:
          type: string
        cultivo:
          type: string
        cantidadKg:
          type: number
        valorUnitario:
          type: number
        total:
          type: number
        notas:
          type: string
      required:
        - id
        - fincaId
        - fecha
        - cultivo
        - cantidadKg
        - valorUnitario
        - total
    CreateProduccionBody:
      type: object
      properties:
        fincaId:
          type: integer
        loteId:
          type: integer
        fecha:
          type: string
        cultivo:
          type: string
        cantidadKg:
          type: number
        valorUnitario:
          type: number
        notas:
          type: string
      required:
        - fincaId
        - fecha
        - cultivo
        - cantidadKg
        - valorUnitario
...
    NominaRecord:
      type: object
      properties:
        id:
          type: integer
        empleadoId:
          type: integer
        empleadoNombre:
          type: string
        periodo:
          type: string
        diasTrabajados:
          type: integer
        salarioBase:
          type: number
        bonificaciones:
          type: number
        deducciones:
          type: number
        totalPagar:
          type: number
        estado:
          type: string
        fechaPago:
          type: string
      required:
        - id
        - empleadoId
        - empleadoNombre
        - periodo
        - diasTrabajados
        - salarioBase
        - totalPagar
        - estado
    CreateNominaBody:
      type: object
      properties:
        empleadoId:
          type: integer
        periodo:
          type: string
        diasTrabajados:
          type: integer
        salarioBase:
          type: number
        bonificaciones:
          type: number
        deducciones:
          type: number
        estado:
          type: string
        fechaPago:
          type: string
      required:
        - empleadoId
        - periodo
        - diasTrabajados
        - salarioBase
        - estado
    Insumo:
      type: object
      properties:
        id:
          type: integer
        fincaId:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        cantidad:
          type: number
        unidad:
          type: string
        costoUnitario:
          type: number
        proveedor:
          type: string
        fechaCompra:
          type: string
        stockMinimo:
          type: number
      required:
        - id
        - nombre
        - tipo
        - cantidad
        - unidad
        - costoUnitario
    CreateInsumoBody:
      type: object
      properties:
        fincaId:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        cantidad:
          type: number
        unidad:
          type: string
        costoUnitario:
          type: number
        proveedor:
          type: string
        fechaCompra:
          type: string
        stockMinimo:
          type: number
      required:
        - nombre
        - tipo
        - cantidad
        - unidad
        - costoUnitario
    Jornada:
      type: object
      properties:
        id:
          type: integer
        empleadoId:
          type: integer
        empleadoNombre:
          type: string
        fincaId:
          type: integer
        fecha:
          type: string
        horasRegulares:
          type: number
        horasExtra:
          type: number
        actividad:
          type: string
        notas:
          type: string
      required:
        - id
        - empleadoId
        - fincaId
        - fecha
        - horasRegulares
        - actividad
    CreateJornadaBody:
      type: object
      properties:
        empleadoId:
          type: integer
        fincaId:
          type: integer
        fecha:
          type: string
        horasRegulares:
          type: number
        horasExtra:
          type: number
        actividad:
          type: string
        notas:
          type: string
      required:
        - empleadoId
        - fincaId
        - fecha
        - horasRegulares
        - actividad
...