# lib/api-zod/src/generated/types/createInsumoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FincaGest360 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInsumoBody {
  fincaId?: number;
  nombre: string;
  tipo: string;
  cantidad: number;
  unidad: string;
  costoUnitario: number;
  proveedor?: string;
  fechaCompra?: string;
  stockMinimo?: number;
}