# lib/api-zod/src/generated/types/createLoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FincaGest360 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLoteBody {
  fincaId: number;
  nombre: string;
  hectareas: number;
  cultivo: string;
  estado: string;
}